yatube/posts/forms.py
from django import forms
from django.contrib.auth import get_user_model

from .models import Post, Comment

User = get_user_model()


class PostForm(forms.ModelForm):
    class Meta:
        model = Post
        fields = ('text', 'group', 'image')
        labels = {
            'text': 'Текст поста',
            'group': 'Группа к которой относится пост'
        }
        help_texts = {
            'text': 'Разместите здесь текст.',
        }


class CommentForm(forms.ModelForm):
    class Meta:
        model = Comment
        fields = ('text',)
        labels = {
            'text': 'Текст поста',
        }
        help_texts = {
            'text': 'Разместите здесь текст.',
        }
